An absolute MESS! Changed spectrum label to vector or string-number pairs; New Instrument class, for use as interface; Hunting bugs in Interpolator!

+++ plugins/Diginstrument/common/Dimension.h
@@ -10,11 +10,11 @@ class Dimension
 {
   public:
     std::string name;
-    int min, max;
+    double min, max;
     bool shifting;
     double defaultValue, currentValue;
 
-    Dimension(std::string name, int min, int max, bool shifting = true) : name(name), min(min), max(max), shifting(shifting), defaultValue(min), currentValue(min) {}
-    Dimension(std::string name, int min, int max, bool shifting, float defaultValue) : name(name), min(min), max(max), shifting(shifting), defaultValue(defaultValue), currentValue(defaultValue) {}
+    Dimension(std::string name, double min, double max, bool shifting = true) : name(name), min(min), max(max), shifting(shifting), defaultValue(min), currentValue(min) {}
+    Dimension(std::string name, double min, double max, bool shifting, float defaultValue) : name(name), min(min), max(max), shifting(shifting), defaultValue(defaultValue), currentValue(defaultValue) {}
 };
 };
